% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ourproj-package.R
\docType{package}
\name{ourproj-package}
\alias{ourproj}
\alias{ourproj-package}
\title{ourproj: Our Datascience Project Workflow}
\description{
Develop reproducible, scalable and sharable data science projects, built on the R packages targets and workflowr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rmgpanw.github.io/ourproj/}
  \item \url{https://github.com/rmgpanw/ourproj}
  \item Report bugs at \url{https://github.com/rmgpanw/ourproj/issues}
}

}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com} (\href{https://orcid.org/0000-0002-0800-2890}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Abraham Olvera Barrios [copyright holder]
}

}
\keyword{internal}
