% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_render_workflowr.R
\name{tar_render_workflowr}
\alias{tar_render_workflowr}
\title{Target factory for workflowr R markdown reports.}
\usage{
tar_render_workflowr(
  include = NULL,
  exclude = NULL,
  include_pattern = NULL,
  exclude_pattern = "^_",
  verbose = FALSE
)
}
\arguments{
\item{include}{Optional character vector of .Rmd files to be rendered.}

\item{exclude}{Optional character vector of .Rmd files to be excluded.}

\item{include_pattern}{By default, include all .Rmd files in \code{workflowr_dir}.}

\item{exclude_pattern}{By default, exclude any files with names prefixed by
"_".}

\item{verbose}{If \code{TRUE}, prints a message listing the .Rmd files for which
targets will be created. Also passed on to \code{\link[workflowr:wflow_build]{workflowr::wflow_build()}}.}
}
\value{
A list of target objects with \code{format = 'file'}, one for each
workflowr report.
}
\description{
Generates a list of target objects to render multiple workflowr R markdown
reports.
}
\examples{
#TODO
}
