% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_render_workflowr.R
\name{tar_render_workflowr_single}
\alias{tar_render_workflowr_single}
\title{Target factory for a single workflowr R markdown report}
\usage{
tar_render_workflowr_single(
  rmd_filename,
  workflowr_dir = "analysis",
  output_dir = "public",
  target_prefix = "WFLWR_"
)
}
\arguments{
\item{rmd_filename}{Name of a workflowr Rmd file to be rendered.}

\item{workflowr_dir}{Directory where workflowr Rmd files are located. By
default this is 'analysis'.}

\item{output_dir}{Directory where html report files will be rendered to. By
default this is 'public'.}

\item{target_prefix}{Prefix for workflowr target names.}
}
\value{
A target object (\code{format = 'file'}) for a single workflowr report.
}
\description{
Helper function for \code{\link[=tar_render_workflowr]{tar_render_workflowr()}}.
}
