% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_track_input_files.R
\name{tar_track_input_files}
\alias{tar_track_input_files}
\title{Target factory for pipeline input files}
\usage{
tar_track_input_files(suffix = "_IN_PATH")
}
\arguments{
\item{suffix}{Suffix indicating global objects that are input file paths.}
}
\value{
A list of target objects with \code{format = 'file'}, one for each input
file.
}
\description{
Searches the local environment for global objects that input file paths and
generates targets for these with \code{format = "file"}.
}
\examples{
# dummy input file to create target for
DATA_IN_PATH <- tempfile(fileext = ".txt")
file.create(DATA_IN_PATH)

# create target for this with `format = "file"`
tar_track_input_files()
}
