% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_track_input_files.R
\name{tar_track_input_files_yaml}
\alias{tar_track_input_files_yaml}
\title{Target factory for pipeline input files}
\usage{
tar_track_input_files_yaml(yaml_file = NULL)
}
\arguments{
\item{yaml_file}{File path to a yaml file}
}
\value{
A list of target objects with \code{format = 'file'}, one for each input
file.
}
\description{
Target factory for pipeline input files
}
\examples{
# TODO
}
