% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_analyses_reactable.R
\name{wflow_analyses_reactable}
\alias{wflow_analyses_reactable}
\title{Create a filterable table of Workflowr analysis reports}
\usage{
wflow_analyses_reactable()
}
\value{
A \url{reactable} table.
}
\description{
To be used in \code{analysis/index.Rmd} within a \href{https://workflowr.github.io/workflowr/index.html}{Workflowr project}.
}
\examples{
\dontrun{
 #TODO
}
}
