% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ourproj_start.R
\name{ourproj_start}
\alias{ourproj_start}
\title{Start a new workflowr/targets project}
\usage{
ourproj_start(directory, project_name, git_username)
}
\arguments{
\item{directory}{character. The directory where the new project will be
created, e.g. "~/myproj". An error is raised if this directory already
exists.}

\item{project_name}{character. The name of the project.}

\item{git_username}{character. Gitlab user name.}
}
\value{
\code{NULL} invisibly.
}
\description{
Creates a directory with the essential files for a workflowr project, using
Gitlab Pages and a targets data analysis pipeline.
}
\examples{
\dontrun{
 ourproj_start(
   directory = tempdir(),
   project_name = "my-proj",
   git_username = "user"
 )
}
}
