% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ourproj_start.R
\name{ourproj_start}
\alias{ourproj_start}
\title{Start a new workflowr/targets project}
\usage{
ourproj_start(
  directory,
  project_title,
  git_username,
  github_gitlab = "gitlab",
  template = "workflowr_targets"
)
}
\arguments{
\item{directory}{character. The directory where the new project will be
created, e.g. "~/myproj". An error is raised if this directory already
exists.}

\item{project_title}{character. The title of the project.}

\item{git_username}{character. GitHub/GitLab user name.}

\item{github_gitlab}{character. Either 'github' or 'gitlab'}

\item{template}{Project template to be used. Options: 'workflowr_targets',
'quarto_website_targets'.}
}
\value{
\code{directory} invisibly.
}
\description{
Creates a directory with the essential files for a
\href{https://workflowr.github.io/workflowr/}{workflowr} or
\href{https://quarto.org/}{quarto} project, using GitHub/Gitlab Pages and a
\href{https://books.ropensci.org/targets/}{targets} data analysis pipeline.
}
\details{
Examples of available project templates:
\itemize{
\item \href{https://rmgpanw.github.io/workflowr_targets/}{workflowr_targets}
\item \href{https://rmgpanw.github.io/quarto_website_targets/}{quarto_website_targets}
}
}
\examples{
\dontrun{
 ourproj_start(
   directory = tempdir(),
   project_title = "My Project",
   git_username = "user"
 )
}
}
